[Refactoring] 파생 Condition 클래스를 Condition.h에 포함 파생 클래스마다 파일을 분리하게 되면 include 문 처리가 복잡해짐
Include문을 효과적으로 관리할 수 있는 방법이 있으면 공유 부탁드립니다.
