--- Source/EmployeeManagement/ConditionBirthday.h
@@ -1,7 +1,0 @@
-#pragma once
-#include "Condition.h"
-
-class ConditionBirthday : public Condition {
-
-
-};
