--- Source/EmployeeManagement/ConditionCL.h
@@ -1,7 +1,0 @@
-#pragma once
-#include "Condition.h"
-
-class ConditionCL : public Condition {
-
-
-};
