--- Source/EmployeeManagement/ConditionCerti.h
@@ -1,7 +1,0 @@
-#pragma once
-#include "Condition.h"
-
-class ConditionCerti : public Condition {
-
-
-};
