--- Source/EmployeeManagement/ConditionDayBirthday.h
@@ -1,7 +1,0 @@
-#pragma once
-#include "Condition.h"
-
-class ConditionDayBirthday : public Condition {
-
-
-};
