--- Source/EmployeeManagement/ConditionFirstName.h
@@ -1,7 +1,0 @@
-#pragma once
-#include "Condition.h"
-
-class ConditionFirstName : public Condition {
-
-
-};
