--- Source/EmployeeManagement/ConditionLastName.h
@@ -1,7 +1,0 @@
-#pragma once
-#include "Condition.h"
-
-class ConditionLastName : public Condition {
-
-
-};
