--- Source/EmployeeManagement/ConditionLastPhoneNumber.h
@@ -1,7 +1,0 @@
-#pragma once
-#include "Condition.h"
-
-class ConditionLastPhoneNumber : public Condition {
-
-
-};
