--- Source/EmployeeManagement/ConditionMiddlePhoneNumber.h
@@ -1,7 +1,0 @@
-#pragma once
-#include "Condition.h"
-
-class ConditionMiddlePhoneNumber : public Condition {
-
-
-};
