--- Source/EmployeeManagement/ConditionMonthBirthday.h
@@ -1,7 +1,0 @@
-#pragma once
-#include "Condition.h"
-
-class ConditionMonthBirthday : public Condition {
-
-
-};
