--- Source/EmployeeManagement/ConditionName.h
@@ -1,7 +1,0 @@
-#pragma once
-#include "Condition.h"
-
-class ConditionName : public Condition {
-
-
-};
