--- Source/EmployeeManagement/ConditionPhoneNumber.h
@@ -1,7 +1,0 @@
-#pragma once
-#include "Condition.h"
-
-class ConditionPhoneNumber : public Condition {
-
-
-};
