--- Source/EmployeeManagement/ConditionYearBirthday.h
@@ -1,7 +1,0 @@
-#pragma once
-#include "Condition.h"
-
-class ConditionYearBirthday : public Condition {
-
-
-};
